+++ source/core/scene/shape/triangle/bvh/triangle_bvh_tree.hpp
@@ -28,7 +28,7 @@ struct Node {
 	uint32_t axis() const;
 	void set_axis(uint32_t axis);
 
-	Children children(uint8_t sign, uint32_t id) const;
+	Children children(int sign, uint32_t id) const;
 
 	bool has_children() const;
 	void set_has_children(bool children);
